requirements.txt
pillow
pyperclip
opencv-python
flask
flask-cors
easygui
colorama